pyproject.toml
[project]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = [
    {name = "«tashka-zu»",email = "«zubkovanatska21@gmail.com»"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.13.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
